round 2 fix
round 2 fix

+++ scripts/fsmedia.py
@@ -362,7 +362,7 @@ class DebugLandmarks(PostProcessAction):  # pylint: disable=too-few-public-metho
             face = detected_face["face"]
             logger.trace("Drawing Landmarks. Frame: '%s'. Face: %s",
                          detected_face["file_location"].parts[-1], idx)
-            aligned_landmarks = face.aligned_landmarks
+            aligned_landmarks = face.aligned_landmarks.astype("int32")
             for (pos_x, pos_y) in aligned_landmarks:
                 cv2.circle(face.aligned_face, (pos_x, pos_y), 2, (0, 0, 255), -1)
